requirements.txt
python-dotenv
pandas
matplotlib
tiktoken
ollama
